doc(Github): 添加 `afdian` 链接
Thank you to everyone, whether you have donated to this project or not. The progress of this project would not be possible without the support of every member of the community, nor without your help in any form — even if you only use this program regularly, your participation is equally valuable and important to the project.